backend/requirements.txt
boto3==1.26.137
fastapi==0.104.1
graphene>=3.0
graphql-core>=3.2.0
python-dotenv==0.19.0
uvicorn==0.24.0
pytest>=8.0.0
